requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
PyMuPDF>=1.23.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0 